new: add logging

+++ src/main/resources/db/migration/V1__Init_database.sql
@@ -30,6 +30,6 @@ create table transaction
     customer_last_name  varchar(30)  null,
     customer_email      varchar(256) null,
     date_time           timestamp    not null,
-    operator_id         bigint       null
+    operator_id         int(11)       null
 );
